backend/watchlist-ms/requirements.txt
fastapi
uvicorn[standard]
kafka-python
cassandra-driver